refactor: 모듈화 및 만든이 폴더 이동
