+++ src/features/main/components/user/index.ts
@@ -1,0 +1,2 @@
+export { default as UserLogin } from './UserLogin';
+export { default as Footer } from './Footer';
