+++ src/pages/index.ts
@@ -15,7 +15,7 @@ import NoticeDetail from '@/pages/main/notice/NoticeDetail';
 import Lost from './main/lost/Lost';
 import LostUpload from './main/lost/LostUpload';
 import LostComplete from './main/lost/LostComplete';
-import MadeBy from './main/notice/MadeBy';
+import MadeBy from './main/made/MadeBy';
 
 export {
   Main,
