+++ src/pages/main/made/index.ts
@@ -1,0 +1,1 @@
+export { default as MadeBy } from './MadeBy';
